+++ app.json
@@ -11,5 +11,6 @@
     "name": "Google Calendar",
     "nameSlug": "google-calendar",
     "classFile": "GoogleCalendar.ts",
-    "description": "Use app to manage your calendar events"
+    "description": "Use app to manage your calendar events",
+     
 }
