+++ helpers/GSGetter.ts
@@ -18,7 +18,7 @@ export class GCGetter {
     private readonly app: GoogleCalendarApp;
 
 
-    public async login(phase: string, logger: ILogger, http: IHttp, modify:IModify, context:SlashCommandContext, req:IApiRequest, end:IApiEndpointInfo): Promise<boolean> {
+    public async login(phase: string, logger: ILogger, http: IHttp, modify:IModify, context:SlashCommandContext): Promise<boolean> {
 
         let signedin: boolean = false;
         
@@ -26,23 +26,19 @@ export class GCGetter {
 
         if (parame == 'auth') {
             //step1,2,3
-            const auth_code = await this.app.getwebhook().stepone(req,end,this.app.getLogger(), http,modify,context);
+          //  const auth_code = await this.app.getwebhook().stepone(req,end,this.app.getLogger(), http,modify,context);
+          const msg = modify.getCreator().startMessage().setSender(context.getSender()).setRoom(context.getRoom());
+          const response = (`${this.urli}client_id=${this.Client_id}&redirect_uri=http://localhost:3000/api/apps/public/c759c4f1-a3c1-4202-8238-c6868633ed87/webhook&scope=https://www.googleapis.com/auth/calendar.readonly&prompt=consent&response_type=code`);
 
-            
-           
-            const newresponse = await http.post(`https://www.googleapis.com/oauth2/v4/token/code=${auth_code}&client_id=${this.Client_id}&client_secret=${this.secret}&redirect_uri=http://localhost:3000/general&grant_type=authorization_code`);
-            
-            if (newresponse.statusCode !== HttpStatusCode.OK || !newresponse.data) {
-              logger.debug('Did not get a valid response', newresponse);
-                throw new Error('Unable to retrieve response with auth code.');
-            }
-
-
-            const acesstoken = new GCResults(newresponse.data);
-
-            if (acesstoken) {
-                signedin = true;
-            }
+          try{
+              msg.setText(response);   
+               await modify.getCreator().finish(msg);
+             }catch (e) {
+                 this.app.getLogger().error('Failed sending login url', e);
+                 msg.setText('An error occurred when trying to send the login url:disappointed_relieved:');
+     
+                 modify.getNotifier().notifyUser(context.getSender(), msg.getMessage());
+             }
 
 
         }
